Add override-parameters subcommand
This command allows a user to create a new change set for an existing
stack, overriding the parameters as provided.

This will never update the underlying stack and is thus the safest
variant to update parameters. This command will only create a change
set, executing it has to be done in a separate step (preferrably using
the regular AWS CLI).