Fix compatibility check to handle unknown commits
Previously if a commit that was part of the compatibility-comparison was
unknown to the local repository, the application exited with a
relatively generic error instead of simply indicating that the compared
changes are incompatible.

We now specifically check if the error returned is of type "bad object",
in which case we mark the changes as incompatible. Unknown commits can
occur if either the deployed changes were created by a different
developer who only had their commit locally, or if a rebase has
occurred.