find_amis_inuse: don't filter launch templates
The `is-default-version` filter apparently is no longer available (or
never was, and just worked "by accident"). Luckily we don't need it
anyway: rather than taking just the latest version of a launch template,
we look into all versions instead.

This also ensures that we don't delete an AMI that might still be
reachable through an old version.

WO-307