update-deployed-template: use template defaults if specified
Previously the defaults specified in a template were not used during
update of a deployed template. This commit changes this, which means if
a default is available on the template, it will be used.

Please note that this default will never overwrite a parameter provided
explicitly through either `--parameters` or `--parameter-path`.

WO-521